issue #28 refactor : entity -> (domain) model mapper 한곳으로 모으기
