+++ client/app/src/main/java/com/example/witt/data/api/auth/response/SignInResponse.kt
@@ -1,15 +1,8 @@
 package com.example.witt.data.api.auth.response
 
-import com.example.witt.domain.model.SignInModel
-
 data class SignInResponse (
     val status : Boolean,
     val reason: String,
     val AccessToken : String,
     val RefreshToken : String
         )
-
-fun SignInResponse.toSignInModel() = SignInModel(
-        status = status,
-        reason = reason
-)
