+++ client/app/src/main/java/com/example/witt/data/api/auth/response/SignUpResponse.kt
@@ -7,7 +7,3 @@ data class SignUpResponse (
     val reason: String
         )
 
-fun SignUpResponse.toSignUpModel() =  SignUpModel(
-        status = status,
-        reason = reason
-)
