+++ client/app/src/main/java/com/example/witt/data/repository/UserRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.example.witt.data.repository
 
 import com.example.witt.data.api.DefaultApiService
 import com.example.witt.data.api.user.request.DuplicateEmailRequest
-import com.example.witt.data.api.user.response.toDuplicateEmailModel
+import com.example.witt.data.mapper.toDuplicateEmailModel
 import com.example.witt.domain.model.user.DuplicateEmailModel
 import com.example.witt.domain.repository.UserRepository
 import javax.inject.Inject
